# hpc-ard-launcher-go/launcher/api/openapi.yaml
openapi: 3.0.0
info:
  description: The Launcher API is the execution layer for the Capsules framework.  It
    handles all the details of launching and monitoring runtime environments.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Launcher API
  version: 3.3.7
servers:
- url: /
security:
- BearerAuth: []
...
paths:
  /info/version/api:
    get:
      operationId: getApiVersion
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
            application/yaml:
              schema:
                type: string
          description: API Version
      security:
      - {}
      summary: Gets the API version
      tags:
      - info
...
  /info/api:
    get:
      operationId: getApi
      responses:
        "200":
          content:
            application/yaml:
              schema:
                additionalProperties:
                  type: object
                type: object
          description: API Definition
      security:
      - {}
      summary: Gets the API definition
      tags:
      - info
...
components:
...
  securitySchemes:
    BearerAuth:
      bearerFormat: A launcher-generated authorization token.
      scheme: Bearer
      type: http